Changed Navbar options & fixed the seperation between sections
The navbar was going over the section headings when you click on the navbar headings, I added the seperation to fix the spacing when you click on the headings to take you that that section. Also got rid of Home since it is only one page and instead added a contact heading and shifted the rest of the headings over to match the overall layout of the website.